Changed the parse routines to prevent too many string duplications.
The semantics is now that parse_desc and parse_xfer take a string and dup it for you. If they return a non-zero result the dup string has gone and your input data structure has been zeroed to make sure no partial parse results are leaked.

A negative return means there was some error in allocating memory, etc. A positive return is the offset in the string at which something was objected to.
In both cases the dup string has been freed and your data structure is zeroed out.

Made both routines __must_checks and fixed the sematics of all callers to consume or free the dup string on return of these functions.