Mmap feature
A new "smb_mmap" command and corresponding driver support to allow
users to mmap SMB pages into virtual memory. A two step mechanism is
required:

1) Issue smb_mmap://smb-name.location[#offset][:extent]
2) Use the "reply=ddd" value quoted in return string as offset argument
   in an mmap call.

The first step - smb_mmap - creates a "ticket" that identifies the page
table to be mapped and returns a ticket identifier that may be used as an
offset argument to mmap.

The driver maintains an internal table of 16 tickets, allocated on-demand.
Entries are removed from the ticket book when the ticket number is used in
an mmap call. If a ticket is not used within [table size] mmap calls, then
it is liable to be revoked. You snooze, you lose.