Sysfs-based mmap
This patch rewrites the smb_mmap feature to create and use an
smb/mmaps subtree to declare the parameters of mmap requests
made upon an SMB, via smb_mmap commands, and to recover and employ
those parameters when servicing mmap calls. This replaces the original
"ticketing" scheme in which details were recorded in internal data
structures. Tickets are now written in sysfs.