xfer_start Framework
This patch presents a basic framework for xfer_start that is based on bind
source and destination signaling of an xfer agent to determine when individual
binds, and thence xfer as a whole, are ready for a DMA transfer to begin.

The signaling portions currently use hack calls for developer convenience.
These will eventually use fabric calls, using doorbell buzzes and stuff like that.

And neither is it yet integrated with the DMA engines: that can happen
once Jimmy's patches are in.