Make driver multi-threaded
Previously fresh work from dev/rddma or the fabric receive was scheduled
as a work item on the shared kernel workqueue. Now that we are recursively
invoking remote operations each request has to run independently.

We are not in Kansas anymore, the driver now must be re-entrant and multi-threaded. Races and bugs may now start to appear in what "was working" code.