Change xfer_<ops> to take a bind instead of a desc.
So for example create_xfer://xfer/dst=src is a third party requestors call
to xfer to create a bind of dst=src. In the distributed fabric calls it
doesn't seem right to just have create_xfer://xfer on its own.

Started to edit the fabric commands to pass the full parameters required
on create_dsts, srcs, etc.