Tidy up events structures and types for use in create_{dsts,srcs}
This should provide the basic types and mechanics we are going to need.
In /sys/rddma there will now be a directory called "events" this is a kset
of named ready events and is created in register_subsys using the
rddma_readies.[ch] type.

Each ready event is a kset with a name, the "action name" given in the
options in create_{dsts,srcs} for either the source or destination. This
is the rddma_events.[ch] type.

Each event in an action is a kobj defined in rddma_event.[ch]. This has
the target address in the desc and the doorbell id. These are derived from
the event_id options given in the reply to create_{srcs,dsts}.