Change options from var[=val] to var[(val)] syntax and improve option handing
I need the syntax change because I need to pass options to specific components
of the bind_params and the x/d=s parsing breaks when x?v=v/d?v=v=s?v=v is used
too many ambiguous uses of "=".

Also ability to add new options to an already parsed object is useful in the
event distribution code so a new function to do that...

Removed the fixed number of option with rest for a variable array of pointers,
stored within the string structure created with dup_str.