[BugFix] smb_delete over fabric
The "smb_delete" command was not working across a fabric.
Turned out to be due to the <name> of the SMB being
included in the fabric request, but not its <location>.

The recipient - which pumps the command into do_operation
like all the rest - had no location context with which to
find the SMB.

Long haul for a short fix.