Fix sync operations. Changes semantics of wait
sync_wait now decrements the count and releases the wait queue when it
decrements to zero. You can't have a send and wait option without racing.

so now send is just like wait, it decrements the count and releases the
waitq if it decrements to zero but doesn't itself wait.

Both send and wait can reset the count if its zero.