Avoid bogus xfer.address in vfi_fabric_dst_ready()
Value of xfer.address was not as assigned when bind was created.
Using the bogus value lead to a fault.

This change applies the logic for resolving the fabric address
from vfi_fabric_src_done() and vfi_fabric_dst_done()
to vfi_fabric_src_ready() and vfi_fabric_dst_ready().