Use thread-unique location for fabric location_find
When sending a location_find request to the fabric before
any local locations exist, we need to create a temporary
vfi_location object to pass to vfi_fabric_call().
Previously, the name of the temporary location was the
same as the location we were trying to find. When we
received the location_find reply, we would delete the
temporary location and create the local shadow for the
fabric location. Since the names were the same, multiple
threads racing to perform location_find on the same name
would collide as one would delete the location which
another was trying to create.

This change creates a temporary location with a name that
is unique to each thread. The name chosen for the temporary
location is the address of the 'uniqueloc' variable on the
thread's stack which is unique while each thread executes
vfi_fabric_loaction_find().